requirements.txt
flask>=3.0
torch>=2.0
torchvision>=0.15
transformers>=4.57
accelerate>=1.0
qwen-vl-utils>=0.0.14
numpy>=1.24
Pillow>=10.0
